feat: 新增信号处理面板，迁移并扩展信号处理功能
- 新增SignalProcessingDock面板，包含信号取反和毛刺过滤功能
- 从DeviceOptionsDock移除毛刺过滤相关代码
- 新增信号取反功能，支持按通道翻转采样位
- 调整侧边栏顺序，新增信号处理侧边栏项
- 优化主题颜色值写法，修复设备像素比返回类型
- 移除滑动抽屉拖拽边距相关冗余代码
- 新增信号处理相关会话保存恢复逻辑

+++ PXView/pv/winnativewidget.cpp
@@ -392,7 +392,7 @@ LRESULT CALLBACK WinNativeWidget::WndProc(HWND hWnd, UINT message, WPARAM wParam
 LRESULT WinNativeWidget::hitTest(HWND hWnd, WPARAM wParam, LPARAM lParam)
 {
     (void)wParam;
-    int k = GetDevicePixelRatio();
+    double k = GetDevicePixelRatio();
     const LONG borderWidth = 4 * k;
     RECT rc;
     GetWindowRect(hWnd, &rc);
@@ -824,7 +824,7 @@ bool WinNativeWidget::IsWin7()
     return false;
 }
 
-int WinNativeWidget::GetDevicePixelRatio()
+double WinNativeWidget::GetDevicePixelRatio()
 {
     auto screen = GetPointScreen();
     if (screen != NULL){
